Add files via upload
To contextualize, the starting point is the complicated situation we are going through in the midst of the COVID-19 pandemic. Today, there is much talk about action plans when the situation is over and how countries and localities will begin to restart their economies. One action that most countries are taking is "contact tracing". 

This has awakened interest in the usefulness of graph databases to solve this type of problem, since it is just as important to be able to identify both direct and indirect contacts.